WIP(server.h): let OFI servers exchange rkey
Note: change rkey from 32bits to 64bits. Might break other backends.

+++ src/include/lci_priv.h
@@ -81,7 +81,7 @@ struct lc_rep {
   void* handle;
   int rank;
   intptr_t base;
-  int32_t rkey;
+  uint64_t rkey;
 };
 
 #include "pool.h"
