+++ src/include/server/server_ibv.h
@@ -216,7 +216,7 @@ static inline void lc_server_sendm(lc_server* s, void* rep, size_t size,
 
 static inline void lc_server_puts(lc_server* s __UNUSED__, void* rep,
                                    void* buf, uintptr_t base, uint32_t offset,
-                                   uint32_t rkey, uint32_t meta, size_t size)
+                                   uint64_t rkey, uint32_t meta, size_t size)
 {
   struct ibv_send_wr this_wr;
   struct ibv_send_wr* bad_wr = 0;
@@ -237,7 +237,7 @@ static inline void lc_server_puts(lc_server* s __UNUSED__, void* rep,
 }
 
 static inline void lc_server_putm(lc_server* s, void* rep, uintptr_t base,
-                                   uint32_t offset, uint32_t rkey, size_t size,
+                                   uint32_t offset, uint64_t rkey, size_t size,
                                    uint32_t meta, lc_packet* ctx)
 {
   struct ibv_send_wr this_wr;
@@ -261,7 +261,7 @@ static inline void lc_server_putm(lc_server* s, void* rep, uintptr_t base,
 
 static inline void lc_server_putl(lc_server* s, void* rep, void* buf,
                                    uintptr_t base, uint32_t offset,
-                                   uint32_t rkey, size_t size, uint32_t meta,
+                                   uint64_t rkey, size_t size, uint32_t meta,
                                    lc_packet* ctx)
 {
   struct ibv_send_wr this_wr;
@@ -286,7 +286,7 @@ static inline void lc_server_putl(lc_server* s, void* rep, void* buf,
 }
 
 static inline void lc_server_rma_rtr(lc_server* s, void* rep, void* buf,
-                                     uintptr_t addr, uint32_t rkey, size_t size,
+                                     uintptr_t addr, uint64_t rkey, size_t size,
                                      uint32_t sid, lc_packet* ctx)
 {
   struct ibv_send_wr this_wr;
